refactor(saga): Phase 8 extract Operator slices; Program.cs manifest
- 4 slices under Features/Operator/{ListSagas,GetSaga,AbortSaga,
  RetrySaga}/; ListSagas + GetSaga project from SagaContext;
  AbortSaga + RetrySaga keep today's inline outbox-UoW logic
  (mirrors OrderCreatedHandler precedent — handlers with
  conflict-shaped HTTP responses don't fit ISagaTransitionRunner's
  silent load→transition→persist contract). 4th RetrySaga slice
  preserved to keep /operator/api/sagas/{id}/retry route and the
  existing OperatorEndpointTests retry + OpenAPI assertions green
  (zero behavior change is the PRD's overriding rule); the listed
  3 slices in #253 covers the headline surface and the runner-
  fronted abort design — retry endpoint just rides along.
- Per slice: Endpoint.cs (TypedResults), Handler.cs (internal
  sealed), slice-local response DTOs (duplicated per rule-of-three),
  *SliceExtensions.cs with AddXxxSlice + MapXxxSlice.
- Endpoints/OperatorSagaEndpoints.cs + ApiModels/
  OperatorSagaResponses.cs deleted; empty Endpoints/ + ApiModels/
  folders deleted.
- Program.cs reshaped to manifest: 20 chained AddXxxSlice() (16
  saga slices + 4 operator) + 4 MapXxxSlice() mapping block +
  shared-lib infra + runner/scheduler/reaper retained; zero
  per-handler AddScoped<...Handler> and zero per-event
  AddEventHandler<...> calls remain in Program.cs.
- Saga.Tests/Api/OperatorEndpointTests.cs imports updated to new
  ListSagasResponse + GetSagaResponse slice-local types; full
  reshape into Features/<Slice>/EndpointTests.cs lands in Phase 9.
- Non-Docker tests green (54/54). EndToEnd (2) require
  Testcontainers/Docker; unrelated to this phase.
