+++ saga-microservice/Saga.Service/Features/Operator/AbortSaga/AbortSagaEndpoint.cs
@@ -1,0 +1,10 @@
+namespace Saga.Service.Features.Operator.AbortSaga;
+
+internal static class AbortSagaEndpoint
+{
+    public static Task<IResult> Handle(
+        Guid id,
+        AbortSagaHandler handler,
+        CancellationToken cancellationToken) =>
+        handler.HandleAsync(id, cancellationToken);
+}
