+++ saga-microservice/Saga.Service/Features/Operator/AbortSaga/AbortSagaResponse.cs
@@ -1,0 +1,7 @@
+namespace Saga.Service.Features.Operator.AbortSaga;
+
+internal sealed record AbortSagaResponse(
+    Guid SagaId,
+    string Status,
+    string CurrentStep,
+    Guid? CommandId);
