+++ saga-microservice/Saga.Service/Features/Operator/RetrySaga/RetrySagaResponse.cs
@@ -1,0 +1,7 @@
+namespace Saga.Service.Features.Operator.RetrySaga;
+
+internal sealed record RetrySagaResponse(
+    Guid SagaId,
+    string Status,
+    string CurrentStep,
+    Guid? CommandId);
